Move Resource interface to web-client
