+++ packages/web-app-files/src/components/Shares/SharedWithMeSection.vue
@@ -98,7 +98,7 @@ import MixinDeclineShare from '../../mixins/actions/declineShare'
 import MixinFilesListFilter from '../../mixins/filesListFilter'
 import MixinMountSideBar from '../../mixins/sidebar/mountSideBar'
 import { useResourcesViewDefaults } from '../../composables'
-import { Resource } from '../../helpers/resource'
+import { Resource } from 'web-client'
 import { useCapabilityShareJailEnabled, useStore } from 'web-pkg/src/composables'
 import { createLocationSpaces } from '../../router'
 import ListInfo from '../../components/FilesList/ListInfo.vue'
