+++ packages/web-app-files/src/helpers/resource/sameResource.ts
@@ -1,4 +1,4 @@
-import { Resource } from './resource'
+import { Resource } from 'web-client'
 
 export const isSameResource = (r1: Resource, r2: Resource): boolean => {
   if (!r1 || !r2) return false
