+++ packages/web-app-files/src/helpers/user/index.ts
@@ -1,2 +1,1 @@
 export * from './avatarUrl'
-export * from './types'
