+++ packages/web-app-files/src/views/Personal.vue
@@ -101,7 +101,8 @@ import ContextActions from '../components/FilesList/ContextActions.vue'
 import { createLocationSpaces } from '../router'
 import { useResourcesViewDefaults } from '../composables'
 import { defineComponent, unref, computed } from '@vue/composition-api'
-import { Resource, move } from '../helpers/resource'
+import { move } from '../helpers/resource'
+import { Resource } from 'web-client'
 import { useGraphClient } from 'web-client/src/composables'
 import { useCapabilityShareJailEnabled, useRouteParam } from 'web-pkg/src/composables'
 import KeyboardActions from '../components/FilesList/KeyboardActions.vue'
