+++ packages/web-app-files/src/views/shares/SharedResource.vue
@@ -104,7 +104,8 @@ import { createLocationSpaces } from '../../router'
 import { useResourcesViewDefaults } from '../../composables'
 import { defineComponent, unref } from '@vue/composition-api'
 import { fetchResources } from '../../services/folder'
-import { Resource, move } from '../../helpers/resource'
+import { move } from '../../helpers/resource'
+import { Resource } from 'web-client'
 import { breadcrumbsFromPath, concatBreadcrumbs } from '../../helpers/breadcrumbs'
 import { useRouteParam, useRouteQuery } from 'web-pkg/src/composables'
 import KeyboardActions from '../../components/FilesList/KeyboardActions.vue'
