+++ packages/web-app-files/src/views/shares/SharedViaLink.vue
@@ -70,7 +70,7 @@ import ContextActions from '../../components/FilesList/ContextActions.vue'
 import { createLocationSpaces } from '../../router'
 import { useResourcesViewDefaults } from '../../composables'
 import { defineComponent } from '@vue/composition-api'
-import { Resource } from '../../helpers/resource'
+import { Resource } from 'web-client'
 import { shareQuickLinkHelp } from '../../helpers/contextualHelpers'
 import { useStore } from 'web-pkg/src/composables'
