+++ packages/web-client/src/index.ts
@@ -5,6 +5,10 @@ import { ocs, OCS } from './ocs'
 export { Graph } from './graph'
 export { OCS } from './ocs'
 
+export * as helpers from './helpers'
+
+export { Resource, User } from './helpers'
+
 interface Client {
   graph: Graph
   ocs: OCS
